+++ src/test/java/org/opensrp/repository/postgres/MhealthEventsRepositoryTest.java
@@ -33,6 +33,8 @@ public class MhealthEventsRepositoryTest extends MhealthBaseRepositoryTest {
 	@Autowired
 	private MhealthClientsRepository mhealthClientsRepository;
 	
+	private static int increment = 1;
+	
 	@BeforeClass
 	public static void bootStrap() {
 		tableNames = Arrays.asList("core.event", "core.event_metadata", "core.client", "core.client_metadata");
@@ -48,7 +50,12 @@ public class MhealthEventsRepositoryTest extends MhealthBaseRepositoryTest {
 	@Override
 	protected Set<String> getDatabaseScriptsAfterExecute() {
 		Set<String> scripts = new HashSet<String>();
-		scripts.add("drop_column.sql");
+		increment = increment + 1;
+		if (increment == 33) {
+			scripts.add("drop_column.sql");
+			return scripts;
+		}
+		
 		return scripts;
 	}
