+++ src/test/java/org/opensrp/repository/postgres/MhealthMigrationRepositoryTest.java
@@ -21,9 +21,11 @@ public class MhealthMigrationRepositoryTest extends MhealthBaseRepositoryTest {
 	@Autowired
 	private MhealthMigrationRepository mhealthMigrationRepository;
 	
+	private static int increment = 1;
+	
 	@BeforeClass
 	public static void bootStrap() {
-		tableNames = Arrays.asList("core.migration,core.event", "core.event_metadata", "core.client",
+		tableNames = Arrays.asList("core.migration", "core.event", "core.event_metadata", "core.client",
 		    "core.client_metadata");
 	}
 	
@@ -37,7 +39,12 @@ public class MhealthMigrationRepositoryTest extends MhealthBaseRepositoryTest {
 	@Override
 	protected Set<String> getDatabaseScriptsAfterExecute() {
 		Set<String> scripts = new HashSet<String>();
-		scripts.add("drop_column.sql");
+		increment = increment + 1;
+		if (increment == 25) {
+			scripts.add("drop_column.sql");
+			return scripts;
+		}
+		
 		return scripts;
 	}
