+++ src/test/java/org/opensrp/service/MhealthEventServiceTest.java
@@ -39,6 +39,8 @@ public class MhealthEventServiceTest extends MhealthBaseRepositoryTest {
 	@Autowired
 	private MhealthClientsRepository mhealthClientsRepository;
 	
+	private static int increment = 1;
+	
 	@BeforeClass
 	public static void bootStrap() {
 		tableNames = Arrays.asList("core.event", "core.event_metadata", "core.client", "core.client_metadata");
@@ -60,7 +62,13 @@ public class MhealthEventServiceTest extends MhealthBaseRepositoryTest {
 	@Override
 	protected Set<String> getDatabaseScriptsAfterExecute() {
 		Set<String> scripts = new HashSet<String>();
-		scripts.add("drop_column.sql");
+		increment = increment + 1;
+		if (increment == 21) {
+			System.err.println("v:" + increment);
+			scripts.add("drop_column.sql");
+			return scripts;
+		}
+		
 		return scripts;
 	}
